Split SimpleLogProcessor and BatchLogProcessor out of single file (#2661)
